split controlapi functional tests into their own package
functional_test.go was also split into one file per resource, to match what we're doing for the RPC handlers too. See #891

I also needed to make some changes to move the tests to a separate package:

* `NewAteletDialer` now takes `options`, and `WithDialCredentials`
lets a test build its own transport credentials. The fake atelet is
reached over insecure transport. Didn't change existing callers. This is the only
non-test change.
