+++ cmd/ateapi/internal/controlapi/dialer.go
@@ -64,10 +64,20 @@ type AteletDialer struct {
 	dialCredentials func(expectedPodUID string) (credentials.TransportCredentials, error)
 }
 
+// DialerOption customizes an AteletDialer built by NewAteletDialer.
+type DialerOption func(*AteletDialer)
+
+// WithDialCredentials overrides how transport credentials are built for a given
+// atelet pod UID. Tests use it to reach a fake atelet over insecure transport
+// while still exercising the real lookup, dial and connection-cache path.
+func WithDialCredentials(build func(expectedPodUID string) (credentials.TransportCredentials, error)) DialerOption {
+	return func(d *AteletDialer) { d.dialCredentials = build }
+}
+
 // NewAteletDialer creates a new AteletDialer. clientBundlePath and serverCAPath
 // are used to build the per-atelet mTLS credentials used for every atelet connection.
-func NewAteletDialer(workerIndexer cache.Indexer, ateletIndexer cache.Indexer, clientBundlePath, serverCAPath string) *AteletDialer {
-	return &AteletDialer{
+func NewAteletDialer(workerIndexer cache.Indexer, ateletIndexer cache.Indexer, clientBundlePath, serverCAPath string, opts ...DialerOption) *AteletDialer {
+	d := &AteletDialer{
 		workerIndexer: workerIndexer,
 		ateletIndexer: ateletIndexer,
 		ateletConns:   lru.New(1024),
@@ -79,6 +89,10 @@ func NewAteletDialer(workerIndexer cache.Indexer, ateletIndexer cache.Indexer, c
 			return credentials.NewTLS(tlsConfig), nil
 		},
 	}
+	for _, opt := range opts {
+		opt(d)
+	}
+	return d
 }
 
 // DialForWorker returns a gRPC connection to the Atelet running on the same node as the specified worker pod.
